Few changes in registrations read controller
1. Remove duplicate controller method from UserProfileController
2. Fix issue in tests with surname claim being used instead of email.
